Add -fno-common to cflags.
This is needed for the reason stated in the comment: I am often declaring
default-initialized global variables as hidden or with weak aliases, neither of
which works if the variable is put in the common section. -fcommon is a hack for
old code, anyway.

I had originally limited the flag to ARM, because my ARM compiler has -fcommon
by default, and I first thought it was ARM specific behavior. But it was not.