Add support for ARM's yield instruction.
Apparently, it does a similar job to pause on x86, namely to deprioritize the
running hardware thread, and I can use it to relax the CPU in spinlock loops. I
only have one spinlock loop.

According to the ARM ARM, this instruction is only supported on ARMv7, though it
will decode as a NOP on ARMv6. But since I will have a special ARMv7 version of
a lot of things coming up anyway, may as well only use it when compiling for
ARMv7.