Enable thumb compilation on ARM.
It turns out that on ARM, the assembler will only emit thumb code if it has been
told. This can be done with -mthumb or with ".code 16". The version with -mthumb
if __thumb2__ is defined in the compiler is probably the best, because it allows
the assembler to select the instruction set dynamically.

I am currently not planning to support Thumb 1. That instruction set is only
used on ARMv4-ARMv6 without ARMv6T2, and it is very limited. The compiler may
generate code for it, but I shan't.