Misc fixes for thumb mode.
When forcing the assembler into thumb mode, it will generate errors if something
cannot be done. Accessing the high registers is generally off limits, and so I
had to make adjustments to the way the stack pointer is initialized.

The function pointers in atomics.S now go through actual symbols. This is needed
to get them to automatically adjust to thumb mode. Although most of that code
only makes sense on Thumb2, so it is possible this was wasted effort.

I also modified the internal calls to __aeabi_memcpy from memcpy (and the same
for memmove) to just use a BL opcode. According to the current AAPCS, the linker
will insert an interworking stub if need be. Also, to facilitate possible use in
Thumb1 (I am still on the fence about it), the ARMv4 code now pops the old LR to
R1 before jumping to it. This is because the Thumb1 version of POP can only
access the low registers and PC, and on ARMv4, POP PC was a simple branch, but I
may need an interworking one.