Force unified syntax for Thumb syscalls.
The lib didn't build for ARMv4T. Still doesn't, because my compiler doesn't
support ARMv4 with VFP, but linking the runtime library requires VFP. Anyway,
GCC already normally uses unified syntax, but switches back to divided for
assembler snippets. I am going to assume that compilers using divided syntax are
few and far between, and for that reason, I only switch the syntax back to
unified at the start. The compiler already switches back to unified at the end.

The reason this is necessary is because the S flag inverts in meaning between
unified and divided syntax, and so "eors r7, r6" in divided means the same as
"eor r7, r6" in unified. However, the latter cannot be represented in Thumb1
mode, which was the whole reason I changed the opcodes to "eors" in the first
place.