Various issues in arm/cpu_arch.h.
LDREX and STREX can only have an offset in Thumb2, and even there it is
limited to the 8-bit range, which I'm not sure how to tell GCC about. So
the constraint needs to be Q and not m, which forces an offset of 0. And
STREX does not permit destination and source registers to be equal,
reqiring an early clobber marking.

I also put the wrong limit on the RBIT instruction. It is only support
on ARMv6T2, not ARMv6 in general.