Add Thumb1 indirect call veneers.
In Armv4 Thumb, the only interworking branch instruction is "bx", which
doesn't set the link register. "blx" could do that, but was only
introduced in ARMv5. To work around the issue, in ARMv4 ARM mode, there
is an equivalent two-instruction sequence. Unfortunately, that sequence
does not exist in Thumb mode (or even worse, it exists but does the
wrong thing). The AAPCS suggests BL'ing to a veneer that does the BX,
but doing that in inline assembler is a bit hacky.

I am doing this with explicit external functions. Their names will be
generated by the compiler appending the register name. It is a hack, but
it works for now.

It is unlikely that the compiler would use a register other than the
four I've used here. If that ever changes, I need to expand the file. At
least that would fail loudly.