Add support for ARM32 TLS descriptors.
ARM32 is the original TLS descriptor architecture, and for that reason
does a few things differently. Its descriptors are backwards, so support
for that needs to be added. For __tlsdesc_dynamic, I debated quite a bit
whether to use __tls_get_addr or inline it, but in the end decided on
inlining for the speed improvements. Apparently, this function can
become hot code, so every instruction and indirection saved is a win.

But for old ARM, there is nothing to be gained from inlining
__aeabi_read_tp. It would just be the exact same code again. So I just
call that function. LR is saved already, anyway.