Fix __a_cas_v[67] retry logic.
I don't always make mistakes, but when I do, I make sure to copy them.
In this case, r0 holds the original expected value only at the start of
the first iteration, and afterwards is 0 if the value was the expected
one (which is the only case when a second iteration is attempted). And I
already had a backup of the original made, just wasn't using it.