Fix ARM instruction set directives.
The build script adds -mthumb to the command line if Thumb2 compilation is
requested. For that reason, all ARM assembler files assemble as ARM or Thumb2
already.

read_tp is special in that I have made it Thumb1 compatible for the case before
ARMv6. It explicitly opts into Thumb1 if __thumb__ is defined before ARMv6.