i386: Add support for __SSE2_MATH__.
It is relatively common to compile i386 code with SSE support. With SSE2, a
second FPU is available, and this one honors the exact type limits given, so
FLT_EVAL_METHOD changes.

musl switches on __FLT_EVAL_METHOD__, but my experiments showed me that only GCC
defines that macro, but not clang. However, both of them define __SSE2_MATH__
when using SSE for FPU, so that is what I am using.