Simplify condition in arm/pthread_arch.h.
Basically, I want to use MRC directly if compile-time arch is at least 6 and we
are compiling for ARM or Thumb2. The previous version divided between v6 and v7
and beyond, but that was not needed. If we are compiling for v6 Thumb1, then we
do need an external function for the instruction switch, but that function can
be __aeabi_read_tp, which in this case will directly read TPIDRURO and return.
So for v6, we never need __a_gettp_ptr.