Optimize thumb syscalls.
Reading the musl source code gave me an idea to depart from the nolibc way of
doing Thumb syscalls. The problem is that r7 may not be available for Thumb
inline assembly. As with i386 before, I therefore have to treat it as never
available, since the conditions are impossible to test. r7 is used as frame
pointer, and is needed if -fno-omit-frame-pointer is present in CFLAGS, or if
optimization is turned off, or if the calling function (these are inline
functions, after all) dynamically allocates stack.

The solution nolibc has for this is to stash the syscall number in r6, then swap
r6 and r7 in the leadup to the syscall, then restore r7 from r6. This solution
works, but it inflexibly puts three unavoidable instructions in front of every
syscall, and allocates a register more, when we have high register pressure in
Thumb1 already.

The musl solution looks somewhat more complicated, but it uses a compound
literal to make the compiler allocate a free register. So not r6 under all
circumstances. Also, the syscall number is not put into the temp register
beforehand, so its value can be lost. In Thumb2 mode, r7 can be set directly
from a literal. In Thumb1, this would work as well, except that GCC has no
constraints to tell it to only use a literal for integers 0-255. Also, the
instruction used would have to be "movs", not "mov".