Add bessel functions from fdlibm.
Downloaded the source code straight from the source. I get to reuse it
in my lib so long as I keep the copyright notice intact, so I'm doing
that.

I chose to split up j0 and y0 into different source files, which means
they need to share pzero() and qzero(), so those become their own hidden
functions. Beyond that, fdlibm just loves to commit aliasing violations,
which I've had to reign in, so __HI() and __LO() were replaced with
__double_bits(), suitably shifted.