Add lgamma() and floor().
I reworked the latter function quite a lot, but the function is mostly
just bit-editing. There are only three cases: Either all bits, some
bits, or no bits of x are fractional. If all bits are fractional, the
only possible answers are 0 and -1. If no bits are fractional, the
answer is the input.

If some bits are fractional, then we have a rounding case. If all
fractional bits are 0 then the input is integral, else they just have to
be removed, optionally adding the mask back in to round away from zero
as necessary.

lgamma() uses that function for some reason. The function was just
copied from fdlibm and siriusified.