Add round-to-int functions.
I decided to remove the copyright notice from floor(), since I have
completely rewritten it. By necessity, I hit the same notes, but the
implementation is my own.

Each of these falls into three cases: If the exponent is negative, then
all representable numbers are fractional, and the only possible outputs
are 0, -0, 1, or -1. Else, if the exponent exceeds the width of the
significand, all representable values already are integers, and can just
be returned as they are. And finally, If the exponent is between those
two values, we need to do things to the fractional bits.

In all cases, if the number returned differs from the number input, an
inexact exception must be raised.

Exception from this is rint(), which simply adds and subtracts 2⁵² in
the right order to turn any input into an integer. However, this is
avoided for the aforementioned already integral inputs to avoid spurious
overflow.

I decided to implement nearbyint() as a wrapper around rint(), because
the other way around would be quite complicated.